Create Scripting101.md
# 302_test_repo
Test repository created during lecture showing how to create a repo
The Python script was created in class to get started on using variables, expressions and statements.
We tried out some statements to pick out syntax and semantic errors (and eventually, runtime errors) in the programmes.
It follows that debugging would become a crucial area of programming.
We covered order of operators,and statement of flow. We used the math module, and noted types of scripts.
We briefly went over keywords used in Python so that they are avoided in naming variables.
Lastly, we "defined" our own functions and "called" on the functions.